Don't check for pumvisible() in CTRL-E
I try to move to the end of the line after a completion often enough
(e.g. after <C-X><C-S><C-N>) that checking for pumvisible() becomes
annoying. If the check is left, one must type <C-Y><C-E> to first accept
the completion and then move to the end of the line.